updated RESTful structure
