--- testforCOMP.py
@@ -1,7 +1,0 @@
-urlist = [1, 2, 3, 4]
-urlist_len = len(urlist)-1
-
-for x in urlist:
-    if urlist.index(x) == urlist_len:
-        print(urlist.index(x), urlist_len)
-
